Extract chargebee license generator
