+++ spec/models/chargebee_license_generator_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe ChargebeeLicenseGenerator do
+end
